refactor last components
