+++ config/locales/en.yml
@@ -522,10 +522,11 @@ en:
     register: Register
     sign_up_hero: Create your account
     sign_up_html: Or %{link}
-  role_level_component:
-    interested_in_roles: Interested in roles
+  role_levels:
+    component:
+      interested_in_roles: Interested in roles
   role_types:
-    role_type_component:
+    component:
       interested_in_roles: Interested in roles
   roles:
     new:
