+++ config/locales/es.yml
@@ -250,10 +250,11 @@ es:
       title: Sin noticaciones leídas
     index:
       title: Ver notificaciones
-  role_level_component:
-    interested_in_roles: Interesado en roles
+  role_levels:
+    component:
+      interested_in_roles: Interesado en roles
   role_types:
-    role_type_component:
+    component:
       interested_in_roles: Interesado en roles
   roles:
     new:
