+++ config/locales/pt-BR.yml
@@ -312,10 +312,11 @@ pt-BR:
     register: Registre-se
     sign_up_hero: Cria sua conta
     sign_up_html: ou %{link}
-  role_level_component:
-    interested_in_roles: Interessado em perfis
+  role_levels:
+    component:
+      interested_in_roles: Interessado em perfis
   role_types:
-    role_type_component:
+    component:
       interested_in_roles: Interessado em perfis
   roles:
     new:
